Account
Added an account file that gets and sets a name and initialises the incoming and outgoing transactions as an empty list. Loops through all of the incoming transactions and adds up the total value of that -same for outgoing transactions. Finally, gets the total amount that the person needs to pay.