Bank file
Added a "bank file" which contained a bank class. Created a list of accounts in "Account" which found all unique names of people in the team, created an empty list of accounts, looped through all the names where for each name, it created the new account and added it to the list. Then returned all the accounts. Another field was made "HashSet" which created a new empty set to hold all the names, then loop through all of the transactions. For each transaction, add the form and to names to the set- the set handles all duplicates and ensures everything is unique. Then returns all the names. In program.cs, used the transactions to create the accounts needed, then returned the accounts.